feat(ui/indexed-tar): inline tar browser on the standalone blob page
Clicking a `.tar` filename (not just the file-list icon) now opens
the same listing UX directly in the page body, so the tar browser
is reachable through both surfaces.

- Extract TarBrowserPanel from TarBrowserDialog. Panel is the full
  body (listing, breadcrumb, hash banner, member view, sub-dialog
  hand-off) without any modal wrapper. Dialog becomes a thin
  el-dialog wrapper around Panel — same props, same behaviour for
  the icon shortcut, just composed.
- Blob page detects indexed-tar inline: when the file is a `.tar`
  whose parent folder also contains `<basename>.json`, fetch the
  tree entry, render <TarBrowserPanel> in place of the binary
  fallback. The fetch is best-effort and silently falls through if
  the sibling is missing or the listing call fails.
- Fix a pre-existing bug in the blob page's repoType detection:
  `path.includes("/models/")` matched in-repo paths like
  `archives/models/bundle.tar`, so the dataset .tar was loading
  under `/api/models/...` and 404-ing. Switch to a first-segment
  check.

Real-browser verification with Playwright drove the actual flow on
the blob page: panel renders, listing shows folders/files, drilling
into weights/router.safetensors and clicking "Open metadata
preview" opens the safetensors metadata sub-dialog with the
seeded tensors and parameter counts.
